Rework CacheWarmer into an inheritable intereface.
Having a cache warmer that works from a file warmer doesn't make sense.
If the file warmer were a warmer, it would have the ability to write
it's contents to the cache.

This reworks CacheWarmer to be simply warmer, which
expects to be inherited by a class that defines `#perform` (similar to
SuckerPunch), instead of `#call` like before.