Update rufus-scheduler
Version 3.0 of rufus-scheduler removes the ::Rufus.to_time_string
method. This commit updates the code to be compatible with that change.
I have not done much testing on this yet, other than getting specs to
pass in a project that is dependent on firebolt. It may be more prudent
to just pin the rufus-scheduler depencency to ~> 2.0, either way works
for me.